+++ src/components/general-widgets/CallMeBack/component.spec.js
@@ -1,15 +1,59 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import { expectComponentToHaveChildren } from '@lodgify/enzyme-jest-expect-helpers/lib/expectComponentToHaveChildren';
+import { expectComponentToHaveProps } from '@lodgify/enzyme-jest-expect-helpers/lib/expectComponentToHaveProps';
 
 import { Form } from 'collections/Form';
+import { SingleDatePicker } from 'inputs/SingleDatePicker';
 import { InputGroup } from 'collections/InputGroup';
 import { Dropdown } from 'inputs/Dropdown';
 
-import * as options from './mock-data/options';
 import { Component as CallMeBack } from './component';
 
-const getCallMeBack = () => shallow(<CallMeBack {...options} />);
+export const propertyOptions = [
+  {
+    text: 'La Casa Viva',
+    value: 'casaViva',
+  },
+  {
+    text: 'La Casa Muerta',
+    value: 'casaMuerta',
+  },
+  {
+    text: 'The White Lodge',
+    value: 'whiteLodge',
+  },
+  {
+    text: 'The Black Lodge',
+    value: 'blackLodge',
+  },
+];
+
+export const timeOptions = [
+  { text: '10 am', value: '1000' },
+  { text: '11 am', value: '1100' },
+  { text: '12 noon', value: '1200' },
+  { text: '1 pm', value: '1300' },
+  { text: '2 pm', value: '1400' },
+  { text: '3 pm', value: '1500' },
+];
+
+export const timeZoneOptions = [
+  { text: 'CET', value: 'cet' },
+  { text: 'GMT', value: 'gmt' },
+  { text: 'EST', value: 'est' },
+];
+
+const getCallMeBack = () =>
+  shallow(
+    <CallMeBack
+      propertyOptions={propertyOptions}
+      timeOptions={timeOptions}
+      timeZoneOptions={timeZoneOptions}
+    />
+  );
 const getForm = () => getCallMeBack().find(Form);
+
 const getFirstInputGroup = () =>
   getCallMeBack()
     .find(InputGroup)
@@ -113,25 +157,19 @@ describe('<CallMeBack />', () => {
   });
 
   describe('the second `InputGroup`', () => {
-    it('should render two `Dropdown`s', () => {
+    it('should render the right children', () => {
       const wrapper = getSecondInputGroup();
-      const actual = wrapper.children(Dropdown);
-      expect(actual).toHaveLength(2);
+      expectComponentToHaveChildren(wrapper, SingleDatePicker, Dropdown);
     });
   });
 
-  describe('the Date `Dropdown`', () => {
+  describe('the `SingleDatePicker`', () => {
     it('should have the right props', () => {
-      const wrapper = getSecondInputGroup();
-      const actual = wrapper.childAt(0).props();
-      expect(actual).toEqual(
-        expect.objectContaining({
-          icon: 'calendar',
-          label: 'Date',
-          name: 'date',
-          options: options.dateOptions,
-        })
-      );
+      const wrapper = getSecondInputGroup().find(SingleDatePicker);
+      expectComponentToHaveProps(wrapper, {
+        placeholderText: 'Date',
+        name: 'date',
+      });
     });
   });
 
@@ -144,7 +182,7 @@ describe('<CallMeBack />', () => {
           icon: 'clock',
           label: 'Time',
           name: 'time',
-          options: options.timeOptions,
+          options: timeOptions,
         })
       );
     });
@@ -166,7 +204,7 @@ describe('<CallMeBack />', () => {
         expect.objectContaining({
           label: 'Time Zone',
           name: 'timeZone',
-          options: options.timeZoneOptions,
+          options: timeZoneOptions,
         })
       );
     });
@@ -180,7 +218,7 @@ describe('<CallMeBack />', () => {
         expect.objectContaining({
           label: 'Property',
           name: 'property',
-          options: options.propertyOptions,
+          options: propertyOptions,
         })
       );
     });